fixed game tied when the team won at last move
Fixed game saying that team has tied when a team clearly wins at the last move (meaning that there is one unfilled tile left and there is a pattern a team can fill and it connects to make a winning pattern). So fixed that by declaring a new variable isGameTied and the function will loop through all the winning patterns if the variable is still true, the game is tied and will display game tied. This will make sure that it displays the result correctly when the game is over.